# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "API Forum App",
        "contact": {
            "email": "rizkimahjati845@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/forum/api",
            "description": "API Forum App"
        }
    ],
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "User - Get All",
                "description": "Mengambil Data User",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil mengambil Data User",
                                    "data": [
                                        {
                                            "id": "1",
                                            "username": "admin",
                                            "password": "admin",
                                            "nama": "Admin",
                                            "alamat": "Pemalang",
                                            "no_telp": "087723977966",
                                            "jenis_kelamin": "Perempuan",
                                            "role": "Admin"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User - Insert",
                "description": "Post data User",
                "operationId": "store",
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "username": "kiki",
                                "password": "password",
                                "nama": "Kiki",
                                "alamat": "Pemalang",
                                "no_telp": "08722729109",
                                "jenis_kelamin": "Perempuan",
                                "role": "user"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Data berhasil disimpan"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/{user}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "User - Get By ID",
                "description": "Mengambil Data User Berdasarkan User ID",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Masukan User ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "3"
                        },
                        "example": "3"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil mengambil Data User",
                                    "data": [
                                        {
                                            "id": "1",
                                            "username": "admin",
                                            "password": "admin",
                                            "nama": "Admin",
                                            "alamat": "Pemalang",
                                            "no_telp": "087723977966",
                                            "jenis_kelamin": "Perempuan",
                                            "role": "Admin",
                                            "api_token": "lalalala",
                                            "token_type": "Bearer"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "User - Update",
                "description": "Update data User",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Masukan User ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "3"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "username": "kiki",
                                "nama": "Kiki",
                                "alamat": "Pemalang",
                                "no_telp": "08772383920",
                                "jenis_kelamin": "Perempuan",
                                "group_id": 3,
                                "role": "User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Data berhasil diubah"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "User - Delete",
                "description": "Delete data User",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Masukan User ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "3"
                        },
                        "example": "2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Data berhasil dihapus"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user-group": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "User - Update Group Id",
                "description": "Update data User",
                "operationId": "updateGroupId",
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "username": "kiki",
                                "group_id": 3
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Data berhasil diubah"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token",
                "scheme": "bearer"
            }
        }
    }
}